@@ -10,6 +10,17 @@ namespace Tlabs.Dynamic {
   ///<summary>Dynamic expression helper</summary>
   public static class DynXHelper {
 
+    ///<summary>Converts <paramref name="methodInfo"/> into a delegeate</summary>
+    public static Delegate AsDelegate(this MethodInfo methodInfo, object? target= null) {
+      var parmTypes= methodInfo.GetParameters().Select(parm => parm.ParameterType);
+      var delegateType= Expression.GetDelegateType(parmTypes.Append(methodInfo.ReturnType).ToArray());
+
+      if (methodInfo.IsStatic)
+        return methodInfo.CreateDelegate(delegateType);
+      return methodInfo.CreateDelegate(delegateType, target);
+    }
+
+
     ///<summary>Compiled context expression info.</summary>
     public struct ContextExpression {
       ///<summary>Context expression</summary>
@@ -74,14 +85,14 @@ namespace Tlabs.Dynamic {
 
       return (new Type[] { type }).Concat(type.GetInterfaces())
                                   .SelectMany(i => i.GetProperties());
-    }    
+    }
 
     ///<summary>
     /// Parse the <paramref name="expression"/> with access to the public properties of <typeparamref name="TCtx"/> (and <paramref name="funcLib"/>) returning <paramref name="retType"/>
     /// into a <see cref="LambdaExpression"/>.
     ///</summary>
     public static LambdaExpression ParsedExpression<TCtx>(string expression, Type retType, IReadOnlyDictionary<string, object> funcLib) {
-      if (null == expression) throw new ArgumentNullException(nameof(expression));
+      ArgumentNullException.ThrowIfNull(expression);
       try {
         return DynamicExpressionParser.ParseLambda(false, typeof(TCtx), retType, expression, funcLib);
       }
