@@ -6,19 +6,19 @@ using System.Threading;
 namespace Tlabs.Misc {
 
   ///<summary>Cache.</summary>
-  public interface ICache<K, T> where T : class {
+  public interface ICache<K, T> where K : notnull where T : class {
 
     ///<summary>Get or set value with <paramref name="key"/>.</summary>
     ///<returns>Cached entry or null if no value cached for <paramref name="key"/>.</returns>
-    T this[K key] { get; set; }
+    T? this[K key] { get; set; }
 
     ///<summary>Gets an already cached value for the given <paramref name="key"/> or if no cached value exists adds the value returned from <paramref name="getValue"/>.</summary>
     ///<returns>The value for the <paramref name="key"/> in the cache or the value returned from <paramref name="getValue"/>.</returns>
-    T this[K key, Func<T> getValue] { get; }
+    T? this[K key, Func<T> getValue] { get; }
 
     ///<summary>Evict entry with <paramref name="key"/>.</summary>
     ///<returns>Evicted entry or null if no value cached for <paramref name="key"/>.</returns>
-    T Evict(K key);
+    T? Evict(K key);
 
     ///<summary>Cached entries.</summary>
     ///<returns>Snapshot of the currently cached entries.</returns>
@@ -26,7 +26,7 @@ namespace Tlabs.Misc {
   }
 
   ///<summary>Basic cache supporting concurrent access with balanced read/write locking.</summary>
-  public class BasicCache<K, T> : IDisposable, ICache<K, T> where T : class {
+  public class BasicCache<K, T> : IDisposable, ICache<K, T> where K : notnull where T : class {
     readonly Dictionary<K, T> cache;
     readonly ReaderWriterLockSlim lck= new();
 
@@ -41,16 +41,17 @@ namespace Tlabs.Misc {
     }
 
     ///<inheritdoc/>
-    public T this[K key] {
+    public T? this[K key] {
       get {
         lck.EnterReadLock();
         try {
           cache.TryGetValue(key, out var val);
-          return val;
+          return val!;
         }
         finally { lck.ExitReadLock(); }
       }
       set {
+        if (null == value) return;
         lck.EnterWriteLock();
         try {
           cache[key]= value;
@@ -72,7 +73,7 @@ namespace Tlabs.Misc {
     }
 
     ///<inheritdoc/>
-    public T Evict(K key) {
+    public T? Evict(K key) {
       lck.EnterUpgradeableReadLock();
       try {
         if (cache.TryGetValue(key, out var val)) {
@@ -111,7 +112,7 @@ namespace Tlabs.Misc {
 
   ///<summary>Cache optimized for concurrent lock free lookups (read).</summary>
   ///<remarks>Concurrent writes are supported but incur more overhead...</remarks>
-  public class LookupCache<K, T> : ICache<K, T> where T : class {
+  public class LookupCache<K, T> : ICache<K, T> where K : notnull where T : class {
     readonly ConcurrentDictionary<K, T> cache;
 
     ///<summary>Default ctor.</summary>
@@ -120,22 +121,22 @@ namespace Tlabs.Misc {
     ///<summary>Default ctor.</summary>
     public LookupCache(IEnumerable<KeyValuePair<K, T>> data) => this.cache= new ConcurrentDictionary<K, T>(data);
     ///<inheritdoc/>
-    public T this[K key] {
+    public T? this[K key] {
       get {
         cache.TryGetValue(key, out var val);
         return val;
       }
-      set { cache[key]= value; }
+      set { if (null != value) cache[key]= value; }
     }
 
     ///<inheritdoc/>
-    public T this[K key, Func<T> getValue] => cache.GetOrAdd(key, (k) => getValue());
+    public T? this[K key, Func<T> getValue] => cache.GetOrAdd(key, (k) => getValue());
 
     ///<inheritdoc/>
     public IEnumerable<KeyValuePair<K, T>> Entries => cache;
 
     ///<inheritdoc/>
-    public T Evict(K key) {
+    public T? Evict(K key) {
       cache.TryRemove(key, out var removedVal);
       return removedVal;
     }
