@@ -4,8 +4,7 @@ using System.Runtime.InteropServices;
 namespace Tlabs.Misc {
   ///<summary>Operating system info</summary>
   public sealed class OSInfo {
-    OSInfo() { } //private ctor
-    internal static event Action<OSPlatform?> OSPlatformResolved; //resolve event for testing
+    internal static event Action<OSPlatform?>? OSPlatformResolved; //resolve event for testing
 
     ///<summary>Current <see cref="OSPlatform"/></summary>
     public static OSPlatform CurrentPlatform => LazyOSInfo.currentOSPlatform;
@@ -16,7 +15,7 @@ namespace Tlabs.Misc {
     }
 
   class LazyOSInfo {
-      static LazyOSInfo() { } //Explicit static ctor for *NOT* to marking type with beforefieldinit
+      static LazyOSInfo() { } //Explicit static ctor for *NOT* to mark type with beforefieldinit
       internal static readonly OSPlatform currentOSPlatform= getCurrentOS() ?? throw new GeneralException($"Unknown paltform: '{Environment.OSVersion}");
       static OSPlatform? getCurrentOS() => Environment.OSVersion.Platform switch {
         PlatformID.Unix     => handlePlatformResolve(RuntimeInformation.IsOSPlatform(OSPlatform.Linux) ? OSPlatform.Linux : null),
