@@ -3,10 +3,14 @@ using Xunit;
 
 namespace Tlabs.Misc.Tests {
 
+  [Collection("App.Setup")]
   public class DateTimeHelperTest {
 
     [Fact]
     public void CurrentTimeTest() {
+      App.Setup= App.Setup with { TimeInfo= Config.ApplicationSetup.Default.TimeInfo };
+      Assert.Equal(DateTimeKind.Utc, App.Setup.TimeInfo.Now.Kind);
+
       var tzid=   System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows)
                ? "W. Europe Standard Time"  //windows
                : "Europe/Berlin";           //non windows
@@ -29,8 +33,8 @@ namespace Tlabs.Misc.Tests {
       Assert.Equal(tInfo.ToUtc(utc).Hour, utc.Hour);
       IConvertible dateTimeSt= "1996-12-19T16:39:57.000000-02:00";
 
-      Assert.Equal(DateTimeKind.Local, dateTimeSt.ToAppTime().Kind);
-      Assert.Equal(19, dateTimeSt.ToAppTime().Hour);
+      Assert.Equal(DateTimeKind.Utc, dateTimeSt.ToAppTime().Kind);
+      Assert.Equal(18, dateTimeSt.ToAppTime().Hour);
     }
   }
 }
