@@ -5,14 +5,14 @@ using Xunit;
 namespace Tlabs.Timing.Test {
   public class TstTimeEnvironment : IDisposable {
     public TstTimeEnvironment() {
-      App.TimeInfo= new DateTimeHelper(TimeZoneInfo.Local);
+      App.Setup= App.Setup with { TimeInfo= new DateTimeHelper(TimeZoneInfo.Local) };
     }
-    public void Dispose() { }
+    public void Dispose() => App.Setup= Config.ApplicationSetup.Default;
   }
 
 
 
-  [Collection("TstTimeScope")]
+  [Collection("App.Setup")]
   public class ScheduleTimeTests {
     private TstTimeEnvironment appTimeEnv;
 
